Moved final methods from Util
